fix linspace needing to be int

+++ simpunch/flow.py
@@ -39,10 +39,10 @@ def generate_flow(gamera_directory: str,
         files_tb = sorted(glob.glob(gamera_directory + "/synthetic_cme/*_TB.fits"))
         files_pb = sorted(glob.glob(gamera_directory + "/synthetic_cme/*_PB.fits"))
 
-        previous_month = np.linspace(1, -30, timedelta(days=30)/time_delta) * time_delta + start_time
+        previous_month = np.linspace(1, -30, int(timedelta(days=30)/time_delta)) * time_delta + start_time
         generate_l3_all_fixed(gamera_directory, previous_month, files_pb[0], files_tb[0])
 
-        next_month = np.linspace(1, 30, timedelta(days=30)/time_delta) * time_delta + start_time
+        next_month = np.linspace(1, 30, int(timedelta(days=30)/time_delta)) * time_delta + start_time
         generate_l3_all_fixed(gamera_directory, next_month, files_pb[-1], files_tb[-1])
 
         # generate_l3_all(gamera_directory, start_time, num_repeats=num_repeats)
